fix: resolve build issue, better package structure
