+++ package.json
@@ -5,7 +5,9 @@
   "packageManager": "pnpm@9.1.3+sha512.7c2ea089e1a6af306409c4fc8c4f0897bdac32b772016196c469d9428f1fe2d5a21daf8ad6512762654ac645b5d9136bb210ec9a00afa8dbc4677843ba362ecd",
   "scripts": {
     "build": "nuxt build",
-    "dev": "pnpm --filter site dev",
+    "dev": "pnpm run \"/dev:/\"",
+    "dev:site": "pnpm --filter site dev",
+    "dev:lib": "pnpm --filter gradient-blob dev",
     "build:lib": "pnpm --filter gradient-blob build",
     "generate": "nuxt generate",
     "preview": "nuxt preview",
