+++ packages/gradient-blob/src/gb/index.ts
@@ -1,0 +1,2 @@
+export * from './cache.js'
+export * from './regular.js'
